requirements.txt
Django~=3.1.3
djangorestframework
markdown
django-filter
djangorestframework_simplejwt